Color-coordination, reorganizing CPU cores, etc...
Each cpu core has a color for the duration of the runtime.

CPU util data was reorganized to make color coordination & data storage simpler.

Commented out prints because they're in the way. I need to add wrap the
printing behind some logging util or output the prints to a logfile.